feat: Re-add --override-provider-settings flag for hybrid mode
Users need the ability to customize provider configurations (e.g., Java heap
size, LSP settings) even when providers run in containers. This flag existed
in the old container mode and should be available in hybrid mode as well.

Changes:
- Added overrideProviderSettings field to analyzeCommand struct
- Added --override-provider-settings CLI flag
- Added validation to ensure file exists if specified

Note: Full implementation to load and merge the override settings file
is needed in a follow-up commit. Currently the flag validates the file
exists but doesn't yet apply the overrides.

Signed-off-by: tsanders <tsanders@redhat.com>